requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.2
tensorflow-cpu>=2.13.0
joblib>=1.3.0
plotly>=5.13.0
matplotlib>=3.7.0
pillow>=9.5.0
keras>=2.13.0